requirements.txt
requests
gradio
json_repair
openai
numpy
transformers
torchaudio
accelerate
